bugs fixINGENEeRING

+++ kjkp.client/src/components/forbiddenWords/Word.js
@@ -9,20 +9,35 @@ var Button = rb.Button;
 var Col = rb.Col;
 var Row = rb.Row;
 
-class Word extends React.Component{
+class Word extends React.Component {
 
-  deleteWord(Id){
-    var r = fetch('http://localhost:8080/zrijeci/delete?Id=' + Id , {
+  deleteWord(Id) {
+
+    var header = {};
+    var auth = AccountService.getAuthInfo();
+
+    if (auth != null && auth.role == 'ROLE_HR') {
+      header = new Headers({
+        'Content-Type': 'application/json; charset=utf8',
+        'Authorization': auth.token
+      });
+
+      var req = makeCancelable(fetch('http://localhost:8080/zrijeci/delete?Id=' + Id, {
         method: 'DELETE',
-        });
-    window.location.reload();
+        headers: header
+      }));
+      req.promise.then(response => response.json())
+        .then(result => window.location.reload())
+        .catch(error => this.setState({ errorMessage: error + "" }));
+      // window.location.reload();
+    }
   }
 
   render() {
     var authenticated = false;
     var role = "";
 
-    var DeleteButton = <a/>;
+    var DeleteButton = <a />;
 
     if (AccountService.getAuthInfo() != null) {
       var auth = AccountService.getAuthInfo();
@@ -30,12 +45,13 @@ class Word extends React.Component{
       authenticated = true;
       role = auth.role;
 
-      if(authenticated && role == "ROLE_HR"){
-        DeleteButton = <Button bsStyle="danger" onClick = {this.deleteWord.bind(this,this.props.word.id)}>Delete</Button>;
+      if (authenticated && role == "ROLE_HR") {
+        DeleteButton = <Button bsStyle="danger" onClick={this.deleteWord.bind(this, this.props.word.id)}>Delete</Button>;
       }
     }
 
-    return(
+
+    return (
       <ListGroupItem>
         <Row>
           <Col xs={6} md={6}>
